Move contact sensor data dataclass to contact sensor implementation.
